Moves compiler to E5R.Tools.Bit.Engine
